fix: POST encounter

+++ src/main/java/fr/alium/dndapi/feature/encounter/entity/dto/EncounterDTO.java
@@ -14,5 +14,5 @@ import java.util.List;
 public class EncounterDTO {
     private String name;
     private String description;
-    private List<Integer> creatures;
+    private List<Long> creatures;
 }
